Abstraindo código em 'routes' e 'controllers'.
