+++ backend/src/server.ts
@@ -1,39 +1,15 @@
 import express from 'express';
-import { getRepository } from 'typeorm';
-import Orphanages from './models/Orphanage';
 
 import './database/connection';
+
+import routes from './routes';
 
 const app = express();
 
 app.use(express.json());
 
-app.post('/orphanages', async (request, response) => {
-  const {
-  name,
-	latitude,
-	longitude,
-	about,
-	instructions,
-	opening_hours,
-	open_on_weekends
-  } = request.body
+app.use(routes);
 
-  const orphanagesRepository = getRepository(Orphanages);
-  
-  const orphanage = orphanagesRepository.create({
-    name,
-    latitude,
-    longitude,
-    about,
-    instructions,
-    opening_hours,
-    open_on_weekends
-    });
 
-    await orphanagesRepository.save(orphanage);
-
-  return response.json({ message: 'Hello world!' })
-});
 
 app.listen(3333);
